Remove Gravatar default on Users.
But also, is this something I need to include in the privacy policy? I am not sharing personally identifiable information with a third party, I am sharing a digest. It only becomes personally identifiable if the user opts-in on Gravatar's side. I'd rather just not deal with this to be sure.